\name{compareestimators}
\alias{compareestimators}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Compare relatedness estimators.}
\description{
Generates simulated individuals of known relatedness, based on observed allele frequencies, then estimates relatedness using 4 methods (lynchli, lynchrd, quellergt, wang) and creates a boxplot of results so that the performance of the different estimators can easily be compared. 
}
\usage{
compareestimators(filename, ninds)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{filename}{
An object generated by the \code{readgenotypedata} function, containing genotype and allele frequency data.
}
  \item{ninds}{
A whole number indicating the number of pairs of individuals of known relatedness you want to generate for each relatedness value (e.g., entering 100 would result in 100 parent-offspring pairs, 100 full-sib pairs, 100 half-sib pairs, and 100 unrelated pairs.)
}
}
\value{
A figure containing box plots comparing the values obtained from each estimator for each known relatedness value.
}
\author{
Tim Frasier <timothy.frasier@smu.ca>
}

\seealso{
\code{\link{readgenotypedata}}, \code{\link{familysim}}, \code{\link{cleanuprvals}}
}
\examples{
	\dontrun{
		#---Load data set---#
		data(GenotypeData)
		input <- readgenotypedata(GenotypeData)

		#---Compare relatedness estimators---#
		compareestimators(input, 100)
	}	
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
