\name{GenotypeData}
\docType{data}
\alias{GenotypeData}
\title{Sample genotype data.}
\description{
Example genotype data containing genotypes of 389 individuals genotyped at 10 codominant loci.  It is the same file that is distributed with the Coancenstry program.
}
\usage{data(GenotypeData)}
\format{
A space-delimited file, where the data for each individual is on a separate line.  Contains one column of individual identifiers.
}
\examples{
data(GenotypeData)
summary(GenotypeData)
}
\keyword{datasets}
