\name{grouprel}
\alias{grouprel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calculate average within-group relatedness and compare to expected values.}
\description{
This function calculates average within-group relatedness and then compares these to expected values for hypothesis testing. Expected values are generated by randomly shuffling individuals between groups, while keeping each group size constant, and calculating average within-group relatedness for these "artificial" groups. Group definition is based on the first two characters of the sample/individual identifiers.
}
\usage{
grouprel(genotypes, estimatorname, usedgroups, iterations)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{genotypes}{
The genotype data to use for the analyses. Note that this is \emph{just} the genotype data, not the entire set of files generated when the data are read into R. Therefore, we have to specify this using the "\code{$gdata}" part of our original file. Can be generated by the \code{\link{familysim}} function.
  }
  \item{estimatorname}{
Tells R what estimator to use. Can be: "dyadml", "lynchli", "lynchrd", "quellergt", "ritland", "tioml", or "wang". Only one estimator can be used at a time.
  }
  \item{usedgroups}{
Indicates what groups to consider in the analyses. This is based on the two-character code that you have used in your individual/sample names. The DEFAULT is "all". However, you can specify specific groups by creating a list of identifiers of the desired groups. For example, with the example file we could select to analyze just the full- and half-sib groups by typing “usedgroups = c("FS", "HS")” here. You can use a similar approach for analyzing only certain groups in your data.
  }
  \item{iterations}{
A single value representing the number of times to shuffle individuals and then recalculate average within-group relatedness for hypothesis testing. 
  }
}
\details{
The function will first estimate all pairwise relatedness values within each group, and take their average. It will print these averages to the screen (but they may pass by too quickly to be seen), and will also write them to a file in R's working directory called "observed-r.csv". This file contains three columns. The first is just a list of integers, indicating how many groups were analyzed, and giving each group an incremental number. The second contains a list of each group "name"", where the name is based on the two-character code from the sample/individual identifiers. These characters will be repeated. For example, the groups are designated "FS", "HS", and "UR" in the example file, so the list here will read "FSFS", "HSHS", and "URUR". This just indicates that the values represent relatedness values between individuals designated "FS" and "FS", "HS" and "HS", and "UR" and "UR".

Next, the function will randomly shuffle the order of individuals in the genotype file, but keep the rows representing each group constant (and the same as the observed data). Thus, individuals are shuffled between groups, while keeping each group size constant. Then relatedness is calculated within each group as described above. This process is repeated however many times is designated in the command.

The results are exported to a file called "expectedrel.csv". This file will contain one column for each group in the analyses, and one row for each iteration performed (the columns are not labeled, however). Thus, each column represents the distribution of expected relatedness values within that group if group membership is random with respect to relatedness. The order of groups (i.e., the order of columns) is the same order as they are encountered in the genotype file (and as printed in the "observed-r.csv" file). The last column is the average relatedness value across all groups for each iteration. This can be used to test global patterns across your data, rather than just for each group independently. By having this as a separate file, you can explore the data on your own.

Figures are automatically generated comparing observed and expected values, as described below.
}
\value{
One figure is generated for each analyzed group. Each figure contains: (1) a histogram of the expected relatedness values within each group; (2) a red arrow indicating where the observed value lies, for easy comparison of observed and expected values; (3) a title indicating the group comparison; and (4) a p-value, indicating the percentage of randomized iterations where the expected values were greater than or equal to the observed value (see below for more explanation). A similar figure is generated for the "overall" data set, where the observed average within-group relatedness across all groups is compared to the expected values. Thus, this function simultaneously assess whether or not relatedness within each group is higher than expected, and whether or not relatedness within all groups is high than expected. P-values cannot be exact in this case, because they are based on simulation rather than direct calculation. Suppose that we conducted 500 simulations, and only had 1 of these simulations with an average within-group relatedness greater than or equal to our observed value. It would be tempting to say that our p-value is 0.002 (1 out of 500). However, we need to be conservative here, because our estimate is based on simulation. The appropriate way to do this is to say that we have observed fewer than 2 out of 500 values less than or equal to our observed value. Thus p < 0.004 (fewer than 2 out of 500). This is how p-values are reported in these figures.

Two files are also generated, and written to R's working directory. One contains the average relatedness within observed groups (called "observed-r.csv"). The other contains the relatedness values for each group for each simulations (described above), called "expectedrel.csv".
}
\author{
Tim Frasier <timothy.frasier@smu.ca>
}
\seealso{
\code{\link{readgenotypedata}}, \code{\link{coancestry}}
}
\examples{
	\dontrun{	
		#---Load Example Data---#
		reldata <- readgenotypedata("simrel.csv")

		#---Conduct analyses---#
		grouprel(genotypes = reldata$gdata, estimatorname = "wang", usedgroups = "all", iterations = 100)
	}	
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

