\name{readrelatedoutput}
\alias{readrelatedoutput}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Read Related output files into R data.frame objects}
\description{
This function is not seen by the user.  It is used for reading the output from the Fortran Coancestry code into appropriate R data frames.
}
\usage{
readrelatedoutput(working.directory = tempdir(), file.prefix = "", any.ci95 = F, 
allow.inbreeding = F)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{working.directory}{
The location of the working directory where Coancestry will place the files.
}
  \item{file.prefix}{
The prefix for the files generated by Coancestry.
}
  \item{any.ci95}{
A TRUE/FALSE statement of whether or not confidence intervals were estimated.
}
  \item{allow.inbreeding}{
A TRUE/FALSE statement of wheter or not inbreeding was allowed/estimated.
}
}

\author{
Jack Pew
}

\seealso{
\code{\link{coancestry}}
}

