\name{simrel}
\docType{data}
\alias{simrel}
\title{Sample genotype data for group relatedness analysis.}
\description{
Example genotype data for group relatedness analysis. Contains genotypes of 30 individuals: 10 full-sibs, 10 half-sibs, and 10 unrelated individuals. The groups are designated by the characters "FS", "HS", and "UR", respectively.
}
\usage{data(simrel)}
\format{
A space-delimited file, where the data for each individual is on a separate line.  Contains one column of individual identifiers.
}
\examples{
data(simrel)
summary(simrel)
}
\keyword{datasets}
